requirements.txt
aiogram==3.13.0
APScheduler==3.10.4
asyncpg==0.29.0
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
